requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Tor proxy support
PySocks>=1.7.1

# User agent rotation
random-user-agent>=1.0.1

# Scheduler
APScheduler>=3.10.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI API (optional - for GPT analysis)
# openai>=1.0.0

# Development/Testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
